Move Entry and EntryProperties to their own file
